requirements.txt
sentence-transformers
torch
torchvision
pillow
qdrant-client
matplotlib
